Move MultiPanelPanel into draw
