+++ samplebase.py
@@ -9,10 +9,8 @@ import argparse
 import asyncio
 import importlib
 import mqtt
-import os
 import pytz
 import sys
-import time
 import types
 from typing import Literal
